Add files via upload
Not sure if this is correct, but take a look. I'm assuming the cost function we are trying to minimize is f(x) := max(ratio(xi)) based on what Grant has made already. This search algorithm takes a free vertex and tries to minimize its ratio by moving in 1 of 6 possible directions. The 7th is just leaving the free vertex where it is.